yegnanet/yegnanet-backend/requirements.txt

azure-functions
azure-cosmos